fix: rename and cleanup tests
