+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["rust-tests"]
+members = ["rattler-build-tests"]
 
 [package]
 name = "rattler-build"
